WaylandServer: Move some legacy code to WaylandServerLegacy header
